\name{commuteplots}
\alias{commuteplots}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{Make Plots of Commute Times
}
\description{Creating boxplots of commute times by day of the week,
as well as creating a histogram of all commute times given.
}
\usage{
commuteplots(commutedf,Leave="Left",Arrive="ArriveDesk",
Time="Total",Group="Day",timebins=6)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{commutedf}{A dataframe of commute times. commutedf may have columns for time left and arrived, or a total time
  spent on road in minutes. A group column may also be included in commutedf, and the default is a character string,
  day of the week each time was collected. Any other columns will be unused by the function.}
  \item{Leave}{Time of departure column name. Default: "Left"}
  \item{Arrive}{Time of arrival column name. Default: "ArriveDesk"}
  \item{Time}{Total commute time column name. Default: "Total"}
  \item{Group}{Group name for analysis by factor. Default: "Day"}
  \item{timebins}{Integer for number of groups to split
  times into for k-means clustering. Default: 6.}
}
\details{All plots are generated by ggplot2 and include generic, appropriate labels.
}
\value{
Nothing is returned, but a boxplot, histogram, and sometimes
scatter plot are generated.
}

\author{
David Burton, URMC Dept. of Biostatistics
}
\note{
wubba lubba dub dub
}

%% ~Make other sections like Warning with \section{Warning }{....} ~


% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{ commute }% use one of  RShowDoc("KEYWORDS")
\keyword{ plots }% __ONLY ONE__ keyword per line
