\name{commuteplots}
\alias{commuteplots}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{Make Plots of Commute Times
}
\description{Creating boxplots of commute times by day of the week, 
as well as creating a histogram of all commute times given. 
}
\usage{
commuteplots(x)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{x}{A dataframe of commute times. x must have a Day column with character strings
  of the day of the week each time was collected, and x must have a Total column containing
  integer values for minutes of commute. Any other columns will be unused by the function.}
}
\details{Both plots are generated by ggplot2 and include generic, appropriate labels. 
}
\value{
Nothing is returned, but a boxplot and histogram are plotted.
}
\references{
%% ~put references to the literature/web site here ~
}
\author{
David Burton
}
\note{
wubba lubba dub dub
}

%% ~Make other sections like Warning with \section{Warning }{....} ~

\seealso{
%% ~~objects to See Also as \code{\link{help}}, ~~~
}
\examples{
##---- Should be DIRECTLY executable !! ----
##-- ==>  Define data, use random,
##--	or do  help(data=index)  for the standard data sets.

## The function is currently defined as
function (x) 
{
  }
}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{ ~kwd1 }% use one of  RShowDoc("KEYWORDS")
\keyword{ ~kwd2 }% __ONLY ONE__ keyword per line
